Comiting all things

+++ Expressions/src/com/william/expressions/AddExpression.java
@@ -47,7 +47,7 @@ public class AddExpression extends Expression {
         right = newRight;
     }
 
-    public int evaluate(int x)
+    public double evaluate(double x)
     {
         return left.evaluate(x) + right.evaluate(x);
     }
@@ -59,4 +59,5 @@ public class AddExpression extends Expression {
         left.print(depth + 1);
         right.print(depth + 1);
     }
+    
 }
